+++ package.json
@@ -11,9 +11,11 @@
     "@types/node": "^14.14.10",
     "@types/react": "^16.9.53",
     "@types/react-dom": "^16.9.8",
+    "@types/react-router-config": "^5.0.1",
     "@types/react-router-dom": "^5.1.6",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
+    "react-router-config": "^5.1.1",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.1",
     "typescript": "~4.0.3",
