+++ src/pages/home/Home.tsx
@@ -1,0 +1,29 @@
+import React from 'react';
+import LoginButton from 'components/LoginButton';
+import LogoutButton from 'components/LogoutButton';
+import Profile from 'components/Profile';
+
+const Home: React.FC = () => {
+  return (
+    <div className="App">
+      <LoginButton />
+      <LogoutButton />
+      <Profile />
+      <header className="App-header">
+        <p>
+          Edit <code>src/App.tsx</code> and save to reload.
+        </p>
+        <a
+          className="App-link"
+          href="https://reactjs.org"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          Learn React
+        </a>
+      </header>
+    </div>
+  );
+};
+
+export default Home;
